Combine DashboardExportPreview and PDFDownloadView These two components have similar code
