+++ packages/lesmis/src/routes/PageRoutes.js
@@ -16,7 +16,7 @@ import { RegisterView } from '../views/RegisterView';
 import { NotAuthorisedView } from '../views/NotAuthorisedView';
 import { VerifyEmailView } from '../views/VerifyEmailView';
 import { ABOUT_PAGE, FQS_PAGE, CONTACT_PAGE } from '../constants';
-import { PDFDownloadView } from '../views/PDFDownloadView';
+import { ExportView, PDF_DOWNLOAD_VIEW } from '../views/ExportView';
 
 const AdminPanel = lazy(() => import('./AdminPanelRoutes'));
 
@@ -79,7 +79,7 @@ export const PageRoutes = React.memo(() => {
           <Footer />
         </Route>
         <Route path={`${path}/pdf-export/:entityCode?`}>
-          <PDFDownloadView />
+          <ExportView viewType={PDF_DOWNLOAD_VIEW} />
         </Route>
         <Route path={`${path}/:entityCode/:view?`}>
           <NavBar />
